==> book/01.prep_collins_glacier.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')

df_met = pd.read_csv('../data/xls_Bulath/Bullath_metadata.csv',sep = '\t')
df_smb      = pd.read_excel('../data/xls_Bulath/Bullath_all.xls', header=None)
smb_data    = df_smb.values[:,1:-1]
smb_time    = df_smb.values[:,0]

df_smb.columns = df_met.Name_orig

smb_coll = []
for i in df_met.Name_orig[1:]:
    subset = (df_smb[['Date', i]])
    subset.rename(columns = {i:'SMB'},  inplace=True)
    subset['Stake'] = i
    g_met = df_met.loc[lambda df_met: df_met['Name_orig'] == i, :]
    subset['Latitude']  = float(g_met.Latitude.values)
    subset['Longitude'] = float(g_met.Longitude.values)
    subset['Elevation'] = float(g_met.Elevation.values)
    subset.dropna(inplace=True)
    date_2 = []
    date_2.append(subset.Date[subset.index[0]])
    for j in subset.Date[0:-1]:
        date_2.append(j)

    smb_2 = []
    smb_2.append(0)
    for j in subset.SMB[0:-1]:
        smb_2.append(j)

    subset['Date1'] = date_2
    subset['SMB1'] = smb_2
    subset['Days'] = (subset['Date'] - subset['Date1']).dt.days
    subset['SMB_diff'] = -(subset['SMB1'] - subset['SMB'])
    subset = subset[['Stake', 'Date', 'Date1','Latitude', 'Longitude',
                     'Elevation', 'Days', 'SMB', 'SMB1', 'SMB_diff']]
    smb_coll.append(subset)

smb_coll_df = pd.concat(smb_coll, ignore_index=True)
smb_coll_df[['SMB_diff', 'SMB']] = smb_coll_df[['SMB_diff', 'SMB']]/100

df_s = smb_coll_df.sort_values(by=['Elevation', 'Date'])

df_s

smb_coll_df1 = df_s[['Stake', 'Date', 'Date1', 'Latitude', 'Longitude', 
                     'Elevation', 'Days','SMB_diff']]

smb_coll_df1.rename(columns={'SMB_diff':'SMB'}, inplace=True)

smb_coll_df1.to_csv('../data/SMB_bellin_diff.csv', index=False, sep='\t')
