==> book/02.preprocessing_input_ERA5.ipynb <==
import sys
import pandas as pd
import dateutil
import numpy as np
sys.path.append('../')
from utilities.era5_down import era5_down
import glob

import warnings
warnings.filterwarnings('ignore')

date_parser = lambda x: dateutil.parser.parse(x, ignoretz=True)
# Improved function to sum dataframe columns which contain nan's
def nansumwrapper(a, **kwargs):
    if np.isnan(a).all():
        return np.nan
    else:
        return np.nansum(a, **kwargs)

#cs_file = '../data/SMB_bellin_diff_1.csv'
#cs_file = '../data/SMB_bellin_season.csv'
cs_file = '../data/SMB_bellin_diff.csv'

dfmb = pd.read_csv(cs_file,
   delimiter='\t', parse_dates=['Date', 'Date1'], na_values='NAN')
labels = dfmb.drop_duplicates(subset=['Stake'])

t2m = 't2m'
Snowfall = 'sf'

files = sorted(glob.glob('../data/ERA_59_20_day/*.nc'))
len(files)

data_pro = []
for st in labels['Stake']:
    stake  = dfmb.loc[lambda df: df['Stake'] == st, :].copy().reset_index()
    df_day = era5_down(files, stake['Longitude'][0],
                       stake['Latitude'][0], 
                       stake['Elevation'][0])
    subset_f0 = []
    for i in range(len(stake)):
        subset = df_day.loc[str(stake.Date1[i]): str(stake.Date[i])].copy()
        subset['PDD'] = subset[t2m].where(subset[t2m] > 0, 0)
        subset['stake'] = stake['Stake'][0]
        subset_f = subset.groupby(['stake']).agg({t2m:np.mean, Snowfall:nansumwrapper, 'PDD':nansumwrapper})
        subset_f.index = ([stake.Date[i]])
        subset_f0.append(subset_f)
    
    data_pro.append(pd.concat([stake, pd.concat(subset_f0).reset_index()], axis=1))

data_inp_four = pd.concat(data_pro)
data_inp_four.drop(['index', 'Date1', 'index'], axis=1, inplace=True)

data_inp_four.columns

data_inp_four

# reset SMB, tp, snowfall and PDD to 0 on 0 days observation
#data_inp_four[['SMB',Prec, Snowfall, 'PDD' ]] = data_inp_four[['SMB',Prec, Snowfall,
#                                                               'PDD' ]].where(data_inp_four['Days'] <= 100,
#                                                                              np.nan)
data_inp_four[['SMB', Snowfall, 'PDD' ]] = data_inp_four[['SMB', Snowfall,'PDD' ]].where(data_inp_four['Days'] != 0, np.nan)

#data_inp_four[['PDD' ]] = data_inp_four[['PDD']].where(data_inp_four['SMB'] < 0, 0)
#data_inp_four[['PDD' ]] = data_inp_four[['PDD']].where(data_inp_four['Elevation'] < 260, 0) # four
#data_inp_four[['PDD' ]] = data_inp_four[['PDD']].where(data_inp_four['Elevation'] < 200, 0) # bell


data_inp_four.dropna(inplace=True)

data_inp_four

data_inp_four.to_csv('../data/SMB_input_belli_ERA5.csv', index=False, sep='\t')
