==> book/06.Correlation_glacier_climate.ipynb <==
# Statistical analysis

## Import libraries

import sys
sys.path.append('../')
from utilities.era5_down import annual_sum, annual_mean, season_df, df_corr_pvalues, sel_period
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')


## Read glaciers data

df1 = pd.read_csv('../data/out/HURD_1960_2020.csv', sep='\t', index_col=['time'], parse_dates=['time'])
df2 = pd.read_csv('../data/out/JOHNSONS_1960_2020.csv', sep='\t', index_col=['time'], parse_dates=['time'])
df3 = pd.read_csv('../data/out/BELLINGSHAUSEN_1960_2020.csv', sep='\t', index_col=['time'], parse_dates=['time'])


## Join SMBs glaciers

df = df1[['SMB']]
df.columns = ['HURD']
df['JOHNSONS'] = df2['SMB']
df['BELLINGSHAUSEN'] = df3['SMB']
df

## Correlations SMBs

df_corr_pvalues(df)

Correlations between SMBs annual glacier and glacier, where significance is indicated in $\textbf{*** (p < 0.001), ** (p < 0.01) and * (p < 0.05)}$.

## Read climate modes data SAM, ENSO, PDO and SOI

df4 = pd.read_csv('../data/climate_mode/climate_modes.csv', sep='\t', index_col=['time'],
parse_dates=['time'])
df4


## Merge climate modes and BELLINGSHAUSEN glacier

df_all = pd.merge(df4, df3, how='left', left_index=True, right_index=True).drop(['HGT'], axis=1)
df_all.to_csv('../data/out/Belling_climate_modes_glacier.csv', sep='\t')
df_all.to_excel('../data/out/Belling_climate_modes_glacier.xlsx', index=True)
df_all

## Monthly correlation

df_all.corr()
df_corr_pvalues(df_all)

## Mean annual correlation

dfmean = annual_mean(df_all[['SAM', 'NINO3.4', 'PDO', 'SOI', 'T2', 'PDD', 'SF', 'SMB', 'MELT', 'Q', 'RZ']])
df_corr_pvalues(dfmean)

## Mean summer correlation

df_DJF = season_df(df_all, 'DJF', 1960, 2020)
df_corr_pvalues(df_DJF)

## Mean fall correlation

df_MAM = season_df(df_all, 'MAM', 1960, 2020)
df_corr_pvalues(df_MAM)

## Mean winter correlation

df_JJA = season_df(df_all, 'JJA', 1960, 2020)
df_corr_pvalues(df_JJA)

## Mean spring correlation

df_SON = season_df(df_all, 'SON', 1960, 2020)
df_corr_pvalues(df_SON)

## Summer period correlation (Dec, Jan, Feb, Mar, May)

df_summer = sel_period(df_all, 'summer', 1960, 2020)
df_corr_pvalues(df_summer)


## Winter period correlation (Jun, Jul, Aug, Set, Oct, Nov)

df_winter = sel_period(df_all, 'winter', 1960, 2020)
df_corr_pvalues(df_winter)
